requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.9

# Data Validation & Models
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
streamlit-calendar>=0.6.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# HTTP & API Clients
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Date & Time
python-dateutil>=2.8.2

# Utilities
loguru>=0.7.0
tenacity>=8.2.3

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
